task2/requirements.txt
nnunetv2==2.4.2
numpy
opencv-python
Pillow
albumentations
seaborn
torch
torchvision
tifffile
SimpleITK
scipy
segmentation-models-pytorch
batchgenerators
monai
scikit-image
